Improve Eslint runner output
This change adds a custom reported to improve the output while 1) showing the progress (so we know it's not stuck) and 2) showing all the warnings (todos) and errors at the end. This was the main motivation because with the previous solution we were not showing the warnings at the end of the lint run.

adeira-source-id: d1ad97f2c25cbe870cd2bfbef4dbba0120d5880c